moving cli logic to it's own file
